fix maxdecimaldigits

+++ MultiPrecision/BigUInt/BigUInt.cs
@@ -9,7 +9,7 @@ namespace MultiPrecision {
         private readonly UInt32[] value;
         public static int Length { get; } = checked(default(N).Value);
         public static int Bits { get; } = checked(Length * UIntUtil.UInt32Bits);
-        public static int MaxDecimalDigits { get; } = checked(Bits * 30103 / 100000 - 4); //10^(4 - 1) = 1000 approx equals 1024
+        public static int MaxDecimalDigits { get; } = checked((int)(Bits * 30103L / 100000L) - 4); //10^(4 - 1) = 1000 approx equals 1024
         public ReadOnlyCollection<UInt32> Value => Array.AsReadOnly(value);
 
         static BigUInt() {
